refactor: extract redundant components and hooks into reusable files
